backend/requirements.txt
Flask==3.0.0
Pillow==10.0.1
python-dotenv==1.0.0
supabase==2.0.0
requests==2.31.0
cors==1.0.1
Flask-CORS==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
